cli/pyproject.toml
[project]
name = "ca3-core"
version = "0.0.36"
description = "ca3 Core is your analytics context builder with the best chat interface."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "ca3 Labs" }]
keywords = ["analytics", "chat", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "cyclopts>=4.4.4",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "rich>=14.0.0",
    "ibis-framework>=9.0.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "jinja2>=3.1.0",
    "apscheduler>=3.10.0",
    "questionary>=2.1.0",
    "posthog>=7.8.0",
    "pyngrok>=7.0.0",
    "python-dateutil>=2.8.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "sqlglot>=26.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
# Database backends
postgres = ["ibis-framework[postgres]>=9.0.0"]
bigquery = ["ibis-framework[bigquery]>=9.0.0", "google-cloud>=0.34.0"]
snowflake = ["ibis-framework[snowflake]>=9.0.0", "snowflake-connector-python[secure-local-storage]>=4.2.0"]
duckdb = ["ibis-framework[duckdb]>=9.0.0"]
clickhouse = ["ibis-framework[clickhouse]>=9.0.0"]
databricks = ["ibis-framework[databricks]>=9.0.0", "certifi>=2024.0.0"]
mysql = ["ibis-framework[mysql]>=9.0.0"]
mssql = ["ibis-framework[mssql]>=9.0.0"]
athena = ["ibis-framework[athena]>=9.0.0"]
trino = ["ibis-framework[trino]>=9.0.0"]
redshift = ["ibis-framework[postgres]>=9.0.0", "sshtunnel>=0.4.0"]
fabric = ["ibis-framework[mssql]>=9.0.0", "azure-identity>=1.19.0"]

# LLM providers
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.76.0"]
mistral = ["mistralai>=1.11.1,<2.0.0"]
gemini = ["google-genai>=1.61.0"]
ollama = ["ollama>=0.4.0"]

# Integrations
notion = ["notion-client>=2.7.0", "notion2md>=2.9.0"]

# Convenience groups
all-databases = [
    "ca3-core[postgres,bigquery,snowflake,duckdb,clickhouse,databricks,mysql,mssql,athena,trino,redshift,fabric]",
]
all-llms = ["ca3-core[openai,anthropic,mistral,gemini,ollama]"]
all = ["ca3-core[all-databases,all-llms,notion]"]
dev = ["pytest-cov", "pytest-timeout>=2.3.0"]

[project.scripts]
ca3 = "ca3_core.main:main"

[project.urls]
Homepage = "https://getca3.io"
Repository = "https://github.com/naolabs/chat"

[dependency-groups]
dev = ["ruff>=0.8.0", "ty>=0.0.9", "ca3-core[all]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
# Include the bin directory with server binary and public assets
artifacts = ["ca3_core/bin/**"]

[tool.hatch.build.targets.wheel]
packages = ["ca3_core"]
exclude = ["**/example/**"]

[tool.hatch.build.targets.sdist]
include = [
    "ca3_core/**",
    "README.md",
    "LICENSE",
]
exclude = ["**/example/**"]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "timeout: mark test with a timeout in seconds (pytest-timeout).",
]
